requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==46.0.3
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==6.0
django-allauth==65.13.1
django-autoslug==1.9.9
django-cache-url==3.4.5
django-cloudinary-storage==0.3.0
django-countries==8.2.0
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-filter==25.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
environs==14.5.0
gunicorn==23.0.0
idna==3.11
marshmallow==4.1.2
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
ruff==0.14.10
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.6.2
whitenoise==6.11.0
